ml_training/pyproject.toml
[tool.poetry]

name = "ml-training"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9, <4"
torch = { version = "^2.1.2", source="pytorch-gpu-src"}
torchvision = { version = "^0.16.2", source="pytorch-gpu-src"}
opencv-python = "^4.9.0.80"
pytorch-lightning = ">=2.0.2,<2.1.0"
wandb = "^0.16.2"
segmentation-models-pytorch = "^0.3.3"
fastapi = {extras = ["all"], version = "^0.109.0"}
celery = "^5.3.6"
python-dotenv = "^1.0.1"
tensorboard = "^2.15.1"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
mlflow = "^2.10.1"
onnx = "^1.15.0"
onnxruntime = "^1.17.0"
albumentations = "^1.3.1"
natsort = "^8.4.0"
pyyaml = "^6.0.1"
ultralytics = "^8.1.9"
onnxsim = "^0.4.35"
easydict = "^1.11"
rasterio = "^1.3.9"
filesplit = "^4.0.1"
shapely = "^2.0.2"
aiofiles = "^23.2.1"
pydantic-settings = "^2.2.1"


[tool.poetry.group.dev.dependencies]
flower = "^2.0.1"
ipykernel = "^6.29.0"


[[tool.poetry.source]]
name = "pytorch-gpu-src"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
